Moved shared lighthouse helpers into its own helper library to reduce code duplication
