+++ lighthouse_provider/lighthouse_back_ends/fake_back_end/pubspec.yaml
@@ -6,7 +6,6 @@ publish_to: none
 environment:
   sdk: '>=2.15.1 <3.0.0'
 
-
 dependencies:
   flutter_web_bluetooth: ^0.0.6+2
   meta: ^1.7.0
@@ -25,3 +24,5 @@ dependencies:
 dev_dependencies:
   lints: ^1.0.0
   test: ^1.16.0
+  lighthouse_test_helper:
+    path: ../../lighthouse_test_helper
