+++ lighthouse_provider/lighthouse_back_ends/fake_back_end/test/ble/fake_device_identifier_test.dart
@@ -1,8 +1,8 @@
 import 'dart:convert';
 
-import 'package:flutter_test/flutter_test.dart';
 import 'package:fake_back_end/fake_back_end.dart';
 import 'package:shared_platform/shared_platform_io.dart';
+import 'package:test/test.dart';
 
 void main() {
   tearDown(() {
