+++ lighthouse_provider/lighthouse_back_ends/fake_back_end/test/fake_ble_back_end_test.dart
@@ -1,13 +1,10 @@
-import 'package:flutter_test/flutter_test.dart';
-import 'package:lighthouse_pm/bloc/lighthouse_v2_bloc.dart';
-import 'package:lighthouse_pm/bloc/vive_base_station_bloc.dart';
 import 'package:fake_back_end/fake_back_end.dart';
 import 'package:lighthouse_provider/lighthouse_provider.dart';
 import 'package:lighthouse_providers/lighthouse_v2_device_provider.dart';
 import 'package:lighthouse_providers/vive_base_station_device_provider.dart';
+import 'package:lighthouse_test_helper/lighthouse_test_helper.dart';
 import 'package:shared_platform/shared_platform_io.dart';
-
-import '../../helpers/fake_bloc.dart';
+import 'package:test/test.dart';
 
 void main() {
   tearDown(() {
@@ -45,7 +42,7 @@ void main() {
   test('Should get Lighthouse device V2', () async {
     SharedPlatform.overridePlatform = PlatformOverride.android;
 
-    final persistence = LighthouseV2Bloc(FakeBloc.normal());
+    final persistence = FakeLighthouseV2Bloc();
 
     final backEnd = FakeBLEBackEnd.instance;
     backEnd.updateLastSeen = (final LHDeviceIdentifier deviceIdentifier) {
@@ -91,7 +88,7 @@ void main() {
   test('Should get Lighthouse device Vive', () async {
     SharedPlatform.overridePlatform = PlatformOverride.android;
 
-    final persistence = ViveBaseStationBloc(FakeBloc.normal());
+    final persistence = FakeViveBaseStationBloc();
 
     final backEnd = FakeBLEBackEnd.instance;
     backEnd.updateLastSeen = (final LHDeviceIdentifier deviceIdentifier) {
