+++ lighthouse_provider/lighthouse_provider/pubspec.yaml
@@ -26,3 +26,5 @@ dev_dependencies:
     path: ../lighthouse_back_ends/fake_back_end
   lighthouse_providers:
     path: ../lighthouse_providers
+  lighthouse_test_helper:
+    path: ../lighthouse_test_helper
