+++ lighthouse_provider/lighthouse_provider/test/lighthouse_provider_test.dart
@@ -4,9 +4,9 @@ import 'package:fake_back_end/fake_back_end.dart';
 import 'package:lighthouse_provider/lighthouse_provider.dart';
 import 'package:lighthouse_providers/lighthouse_v2_device_provider.dart';
 import 'package:lighthouse_providers/vive_base_station_device_provider.dart';
+import 'package:lighthouse_test_helper/lighthouse_test_helper.dart';
 import 'package:test/test.dart';
 
-import 'helpers/fake_bloc.dart';
 import 'lighthouse_back_ends/fake/fake_pair_back_end.dart';
 
 void main() {
