+++ lighthouse_provider/lighthouse_providers/test/vive_base_station/device/vive_base_station_device_test.dart
@@ -2,12 +2,9 @@ import 'package:fake_back_end/fake_back_end.dart';
 import 'package:lighthouse_provider/device_extensions/device_extension.dart';
 import 'package:lighthouse_provider/lighthouse_provider.dart';
 import 'package:lighthouse_providers/vive_base_station_device_provider.dart';
+import 'package:lighthouse_test_helper/lighthouse_test_helper.dart';
 import 'package:shared_platform/shared_platform_io.dart';
 import 'package:test/test.dart';
-
-import '../../helpers/device_helpers.dart';
-import '../../helpers/failing_ble_device.dart';
-import '../../helpers/fake_bloc.dart';
 
 void main() {
   setUp(() {
