+++ lighthouse_provider/lighthouse_test_helper/CHANGELOG.md
@@ -1,0 +1,3 @@
+## 1.0.0
+
+- Created package to reduce code duplication.
