+++ lighthouse_provider/lighthouse_test_helper/README.md
@@ -1,0 +1,5 @@
+# Lighthouse test helper
+
+Some shared test methods used for testing the lighthouse provider.
+
+This is mostly here to reduce duplicated code.
