+++ lighthouse_provider/lighthouse_test_helper/lib/lighthouse_test_helper.dart
@@ -1,0 +1,11 @@
+///
+/// A library with some helper functions to help with testing
+///
+library lighthouse_test_helper;
+
+export 'src/device_helpers.dart';
+export 'src/failing_ble_device.dart';
+export 'src/fake_ble_device_provider.dart';
+export 'src/fake_high_level_device.dart';
+export 'src/fake_persistence.dart';
+export 'src/fake_characteristic.dart';
