+++ lighthouse_provider/lighthouse_test_helper/lib/src/fake_characteristic.dart
@@ -1,12 +1,15 @@
 import 'package:fake_back_end/fake_back_end.dart';
 import 'package:lighthouse_provider/lighthouse_provider.dart';
+import 'package:meta/meta.dart';
 
+@visibleForTesting
 class TestReadCharacteristic extends FakeReadOnlyCharacteristic {
   TestReadCharacteristic(final List<int> data)
       : super(data,
             LighthouseGuid.fromString('FFFFFFFF-0000-0000-0000-000000000000'));
 }
 
+@visibleForTesting
 class TestReadWriteCharacteristic extends FakeReadWriteCharacteristic {
   TestReadWriteCharacteristic()
       : super(
