+++ lighthouse_provider/lighthouse_test_helper/pubspec.yaml
@@ -1,0 +1,26 @@
+name: lighthouse_test_helper
+description: A starting point for Dart libraries or applications.
+version: 1.0.0
+publish_to: none
+
+environment:
+  sdk: '>=2.15.1 <3.0.0'
+
+
+dependencies:
+  meta: ^1.7.0
+  rxdart: ^0.27.3
+  shared_platform:
+    path: ../../shared/shared_platform
+  fake_back_end:
+    path: ../lighthouse_back_ends/fake_back_end
+  lighthouse_back_end:
+    path: ../lighthouse_back_end
+  lighthouse_provider:
+    path: ../lighthouse_provider
+  lighthouse_providers:
+    path: ../lighthouse_providers
+  test: ^1.16.0
+
+dev_dependencies:
+  lints: ^1.0.0
