+++ pubspec.yaml
@@ -91,6 +91,8 @@ dev_dependencies:
   drift_dev: ^1.2.0
   build_runner: 2.1.7
   flutter_lints: ^1.0.4
+  lighthouse_test_helper:
+    path: lighthouse_provider/lighthouse_test_helper
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
